refactor file structor
