+++ package.json
@@ -101,6 +101,7 @@
     "@types/ws": "^4.0.1",
     "bl": "^1.2.0",
     "brolog": "^1.2.0",
+    "clone-class": "^0.4.1",
     "cuid": "^2.1.1",
     "hot-import": "^0.1.0",
     "mime": "^2.2.0",
