+++ src/puppet-web/puppet-web.spec.ts
