+++ src/wechaty.spec.ts
